refactor: cache 적용 후 none-fetch로 metaData 불러오기
