+++ app/posts/[postId]/page.tsx
@@ -3,8 +3,9 @@ import HashTag from "./components/hashTag";
 import SeriesBoard from "app/components/clientside/seriesBoard";
 import styles from "./styles/page.module.scss";
 import { CommentBox } from "./components/clientside";
-import { getPost, mdParser } from "./utils";
+import { getPost, mdParser, getMetaData } from "./utils";
 import { CardLayout } from "app/components/card";
+import { Metadata } from "next/types";
 
 export default async function Posts({
   params,
@@ -68,4 +69,35 @@ export async function generateStaticParams() {
   }
 
   return staticData;
+}
+
+export async function generateMetadata({
+  params,
+}: {
+  params: { postId: string };
+}): Promise<Metadata> {
+  const postId = params.postId;
+  const metaData = await getMetaData(postId);
+
+  return {
+    title: metaData.title,
+    description: metaData.description,
+    keywords: metaData.hashtag,
+    bookmarks: [`${URL}/posts/${postId}`],
+    openGraph: {
+      title: metaData.title,
+      description: metaData.description,
+      url: `${URL}/posts/${postId}`,
+      siteName: "ChocoHam 개발 블로그",
+      images: [{ url: metaData.thumbnail, width: 380, height: 250 }],
+      type: "website",
+    },
+    twitter: {
+      card: "summary_large_image",
+      title: metaData.title,
+      description: metaData.description,
+      creator: "초코햄",
+      images: [metaData.thumbnail],
+    },
+  };
 }
