+++ app/posts/[postId]/utils/index.ts
@@ -1,13 +1,15 @@
 import getPost from "./getPost";
 import commentHandler from "./commentHandler";
 import mdParser from "./mdParser";
+import getMetaData from "./getMetaData";
 
 const Components = {
   getPost,
   commentHandler,
   mdParser,
+  getMetaData,
 };
 
 export default Components;
 
-export { getPost, commentHandler, mdParser };
+export { getPost, commentHandler, mdParser, getMetaData };
